requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==23.2.1
    # via gcloud-aio-storage
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.10.11
    # via gcloud-aio-auth
aioredis==2.0.1
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
anyio==4.5.2
    # via
    #   httpcore
    #   starlette
async-timeout==5.0.1
    # via
    #   aiohttp
    #   aioredis
    #   redis
attrs==25.1.0
    # via aiohttp
backoff==2.2.0
    # via
    #   -r requirements.in
    #   gcloud-aio-auth
    #   opentelemetry-exporter-otlp-proto-http
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via gcloud-aio-auth
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
cryptography==44.0.1
    # via gcloud-aio-auth
deprecated==1.2.18
    # via opentelemetry-api
environs==9.5.0
    # via
    #   -r requirements.in
    #   gundi-client
    #   gundi-client-v2
exceptiongroup==1.2.2
    # via anyio
fastapi==0.108.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gcloud-aio-auth==5.3.2
    # via
    #   gcloud-aio-pubsub
    #   gcloud-aio-storage
gcloud-aio-pubsub==6.0.0
    # via -r requirements.in
gcloud-aio-storage==9.2.0
    # via -r requirements.in
google-api-core[grpc]==2.24.1
    # via google-cloud-trace
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-trace
google-cloud-trace==1.16.0
    # via opentelemetry-exporter-gcp-trace
googleapis-common-protos==1.68.0
    # via
    #   google-api-core
    #   grpcio-status
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
gundi-client==1.0.4
    # via -r requirements.in
gundi-client-v2==2.3.8
    # via -r requirements.in
gundi-core==1.10.2
    # via
    #   -r requirements.in
    #   gundi-client
    #   gundi-client-v2
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==2.3.2
    # via -r requirements.in
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via
    #   -r requirements.in
    #   gundi-client
    #   gundi-client-v2
    #   respx
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
marshmallow==3.22.0
    # via environs
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
opentelemetry-api==1.14.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-propagator-gcp
    #   opentelemetry-sdk
opentelemetry-exporter-gcp-trace==1.3.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-http==1.15.0
    # via -r requirements.in
opentelemetry-instrumentation==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-aiohttp-client==0.35b0
    # via -r requirements.in
opentelemetry-instrumentation-httpx==0.35b0
    # via -r requirements.in
opentelemetry-propagator-gcp==1.9.0
    # via -r requirements.in
opentelemetry-proto==1.15.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.14.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
opentelemetry-util-http==0.35b0
    # via opentelemetry-instrumentation-aiohttp-client
packaging==24.2
    # via marshmallow
prometheus-client==0.21.1
    # via gcloud-aio-pubsub
propcache==0.2.0
    # via yarl
proto-plus==1.26.0
    # via
    #   google-api-core
    #   google-cloud-trace
protobuf==4.25.6
    # via
    #   google-api-core
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpcio-status
    #   opentelemetry-proto
    #   proto-plus
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   gcloud-aio-storage
    #   google-auth
pycparser==2.22
    # via cffi
pydantic==1.10.15
    # via
    #   -r requirements.in
    #   fastapi
    #   gundi-client
    #   gundi-client-v2
    #   gundi-core
pyjwt==2.9.0
    # via gcloud-aio-auth
python-dotenv==1.0.1
    # via environs
redis==5.2.1
    # via walrus
requests==2.32.3
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-http
respx==0.21.1
    # via
    #   gundi-client
    #   gundi-client-v2
rsa==4.9
    # via
    #   gcloud-aio-storage
    #   google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.32.0.post1
    # via fastapi
typing-extensions==4.12.2
    # via
    #   aioredis
    #   anyio
    #   fastapi
    #   multidict
    #   opentelemetry-sdk
    #   pydantic
    #   uvicorn
urllib3==2.2.3
    # via requests
uvicorn==0.25.0
    # via -r requirements.in
walrus==0.9.2
    # via -r requirements.in
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
yarl==1.15.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
